Move rolling impl details to detail/ directory.
Fixes #11211.

The contents of `cpp/src/rolling/` are a mix of `_detail` files that are private to implementation. This includes `rolling_detail.hpp`, `rolling_detail.cuh`, and an assortment of aggregation-specific files.
This commit moves these under `cpp/src/rolling/detail`, thus bringing `rolling` into compliance with other modules.
