+++ cpp/src/rolling/detail/range_window_bounds.hpp
