+++ cpp/src/rolling/detail/rolling.hpp
